backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==9.5.0
django-allauth==0.57.2
requests==2.31.0

gunicorn==21.2.0
django-redis==5.4.0
reportlab==4.4.10
qrcode==8.0.0
